:white_check_mark: Rename test classes to follow naming conventions
