+++ composeApp/src/desktopTest/kotlin/com/crosspaste/app/AppWindowManagerTest.kt
@@ -5,7 +5,7 @@ import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertNull
 
-class TestAppWindowManager {
+class AppWindowManagerTest {
 
     @Test
     fun testMockTestAppWindowManager() {
